feat #21 : 크롤러, API endpoint 연결

+++ src/main/java/com/inha/capstone/controller/ApplicationController.java
@@ -2,14 +2,18 @@ package com.inha.capstone.controller;
 
 import com.inha.capstone.Dto.ApplicationDto;
 import com.inha.capstone.Dto.ApplicationDto.*;
+import com.inha.capstone.config.BaseException;
 import com.inha.capstone.config.BaseResponse;
+import com.inha.capstone.config.BaseResponseStatus;
 import com.inha.capstone.domain.Application;
 import com.inha.capstone.domain.ApplicationCategory;
+import com.inha.capstone.domain.FileCategory;
 import com.inha.capstone.domain.User;
 import com.inha.capstone.service.ApplicationService;
 import com.inha.capstone.service.FileService;
 import com.inha.capstone.service.UserService;
 import com.inha.capstone.util.ApplicationUtil;
+import com.inha.capstone.util.Crawler;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -19,6 +23,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.security.Principal;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -37,6 +42,7 @@ public class ApplicationController {
     private final ApplicationService applicationService;
     private final UserService userService;
     private final FileService fileService;
+    private final Crawler crawler;
 
     @PostMapping("/v1/application")
     public ResponseEntity<BaseResponse<Void>> createApplicationV1(Principal principal, @RequestBody @Valid CreateApplicationRequestV1 request) throws ParseException {
@@ -58,28 +64,42 @@ public class ApplicationController {
     }
 
     @PostMapping("/application")
-    public ResponseEntity<BaseResponse<Void>> createApplication(
+    public ResponseEntity<BaseResponse<CreateApplicationResponse>> createApplication(
         Principal principal,
-        @RequestPart(value = "file") MultipartFile multipartFile,
+        @RequestPart(value = "uiImage") MultipartFile multipartUiFile,
+        @RequestPart(value = "thumbnailImage", required = false) MultipartFile multipartThumbnailFile,
         @RequestPart(value = "info") CreateApplicationRequest request
     ) throws ParseException {
         String endpointPath = "/application";
         log.info(getRequestLog(POST, endpointPath, request));
 
         User user = userService.findByUserId(principal.getName());
-        String uiImagePath = fileService.saveFileToLocal(multipartFile);
-        // TODO: uiImagePath 로 이미지를 html, css 코드로 변환후 해당 파일들이 존재하는 directoryPath 를 uiFilePath 에 저장
-        String uiFilePath = null;
         Application application = new Application(request, user);
         applicationService.save(application);
-        String applicationUrl = fileService.uploadApplicationToS3(uiFilePath,application.getApplicationId(),user.getId());
+
+        String uiFilePath = null;
+        try {
+            uiFilePath = crawler.makeHtml(multipartUiFile);
+        } catch (IOException e) {
+            fileService.deleteFileFromLocal(uiFilePath);
+            throw new BaseException(BaseResponseStatus.CRAWLER_ERROR);
+        } catch (InterruptedException e) {
+            fileService.deleteFileFromLocal(uiFilePath);
+            throw new BaseException(BaseResponseStatus.CRAWLER_ERROR);
+        }
+        String applicationUrl = fileService.uploadFileToS3WithoutDatabase(FileCategory.APPLICATION, uiFilePath, application.getApplicationId(), user.getId());
         application.setApplicationUrl(applicationUrl);
-        applicationService.save(application);
-        fileService.deleteFileFromLocal(uiImagePath);
         fileService.deleteFileFromLocal(uiFilePath);
 
-        log.info(getResponseLog(POST, endpointPath, request, null));
-        return ResponseEntity.ok().body(new BaseResponse());
+        if(multipartThumbnailFile != null) {
+            String thumbnailUrl = fileService.uploadFileToS3(multipartThumbnailFile, FileCategory.IMAGE, application.getApplicationId(), user.getId(), 0L);
+            application.setThumbnailUrl(thumbnailUrl);
+        }
+        application = applicationService.save(application);
+        CreateApplicationResponse response = new CreateApplicationResponse(application.getApplicationUrl());
+
+        log.info(getResponseLog(POST, endpointPath, request, response));
+        return ResponseEntity.ok().body(new BaseResponse(response));
     }
 
     @GetMapping("/v1/applications/{applicationId}")
@@ -113,9 +133,10 @@ public class ApplicationController {
     }
 
     @PatchMapping("/applications/{applicationId}")
-    public ResponseEntity<BaseResponse<Void>> updateApplication(
+    public ResponseEntity<BaseResponse<UpdateApplicationResponse>> updateApplication(
         Principal principal,
-        @RequestPart(value = "uiImage") MultipartFile multipartFile,
+        @RequestPart(value = "uiImage", required = false) MultipartFile multipartUiFile,
+        @RequestPart(value = "thumbnailImage", required = false) MultipartFile multipartThumbnailFile,
         @RequestPart(value = "info") UpdateApplicationRequest request,
         @PathVariable Long applicationId
     ) throws ParseException{
@@ -125,17 +146,33 @@ public class ApplicationController {
 
         User user = userService.findByUserId(principal.getName());
         Application application = applicationService.findById(applicationId);
-        String uiImagePath = fileService.saveFileToLocal(multipartFile);
         // TODO: uiImagePath 로 이미지를 html, css 코드로 변환후 해당 파일들이 존재하는 directoryPath 를 uiFilePath 에 저장
-        String uiFilePath = null;
-        String applicationUrl = fileService.uploadApplicationToS3(uiFilePath, applicationId, user.getId());
-        application.updateApplication(request);
-        applicationService.save(application);
-        fileService.deleteFileFromLocal(uiImagePath);
-        fileService.deleteFileFromLocal(uiFilePath);
 
-        log.info(getResponseLog(PATCH, endpointPath, request, null));
-        return ResponseEntity.ok().body(new BaseResponse());
+        if(multipartUiFile != null) {
+            String uiFilePath = null;
+            try {
+                uiFilePath = crawler.makeHtml(multipartUiFile);
+            } catch (IOException e) {
+                fileService.deleteFileFromLocal(uiFilePath);
+                throw new BaseException(BaseResponseStatus.CRAWLER_ERROR);
+            } catch (InterruptedException e) {
+                fileService.deleteFileFromLocal(uiFilePath);
+                throw new BaseException(BaseResponseStatus.CRAWLER_ERROR);
+            }
+            String applicationUrl = fileService.uploadFileToS3WithoutDatabase(FileCategory.APPLICATION, uiFilePath, applicationId, user.getId());
+            application.setApplicationUrl(applicationUrl);
+            fileService.deleteFileFromLocal(uiFilePath);
+        }
+        if(multipartThumbnailFile != null) {
+            String thumbnailUrl = fileService.uploadFileToS3(multipartThumbnailFile, FileCategory.IMAGE, applicationId, user.getId(), 0L);
+            application.setThumbnailUrl(thumbnailUrl);
+        }
+        application.updateApplication(request);
+        application = applicationService.save(application);
+        UpdateApplicationResponse response = new UpdateApplicationResponse(application.getApplicationUrl());
+
+        log.info(getResponseLog(PATCH, endpointPath, request, response));
+        return ResponseEntity.ok().body(new BaseResponse(response));
     }
 
     @GetMapping("/test/{applicationId}")
